refactor(util): 클래스명 변경 (PurchaseAmountValidator -> PurchaseAmountParser)
- 클래스의 실제 역할이 문자열을 파싱해서 구매 금액을 long 타입으로 변환하는데 가까워 이름 변경
- validate -> parse 메서드명 변경
