+++ src/test/java/lotto/validator/PurchaseAmountParserTest.java
@@ -1,5 +1,6 @@
 package lotto.validator;
 
+import lotto.util.PurchaseAmountParser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -8,13 +9,13 @@ import org.junit.jupiter.params.provider.ValueSource;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.Assertions.assertThat;
 
-class PurchaseAmountValidatorTest {
+class PurchaseAmountParserTest {
 
     @ParameterizedTest
     @ValueSource(strings = {"", " ", "   "})
     void 금액이_비어있으면_예외가_발생한다(String input) {
         // when & then
-        assertThatThrownBy(() -> PurchaseAmountValidator.validate(input))
+        assertThatThrownBy(() -> PurchaseAmountParser.parse(input))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage(ErrorMessage.EMPTY_AMOUNT.getMessage());
     }
@@ -23,7 +24,7 @@ class PurchaseAmountValidatorTest {
     @ValueSource(strings = {"abc", "1000a", "1_000"})
     void 금액이_숫자가_아니면_예외가_발생한다(String input) {
         // when & then
-        assertThatThrownBy(() -> PurchaseAmountValidator.validate(input))
+        assertThatThrownBy(() -> PurchaseAmountParser.parse(input))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage(ErrorMessage.INVALID_NUMBER.getMessage());
     }
@@ -33,7 +34,7 @@ class PurchaseAmountValidatorTest {
     void long_범위를_초과한_입력은_예외가_발생한다() {
         // long 최대값 + 1 입력
         String tooLarge = "9223372036854775808"; // Long.MAX_VALUE + 1
-        assertThatThrownBy(() -> PurchaseAmountValidator.validate(tooLarge))
+        assertThatThrownBy(() -> PurchaseAmountParser.parse(tooLarge))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage(ErrorMessage.OUT_OF_RANGE.getMessage(Long.MIN_VALUE, Long.MAX_VALUE));
     }
@@ -41,7 +42,7 @@ class PurchaseAmountValidatorTest {
     @Test
     void 금액이_0이면_예외가_발생한다() {
         // when & then
-        assertThatThrownBy(() -> PurchaseAmountValidator.validate("0"))
+        assertThatThrownBy(() -> PurchaseAmountParser.parse("0"))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage(ErrorMessage.NEGATIVE_NUMBER.getMessage());
     }
@@ -49,7 +50,7 @@ class PurchaseAmountValidatorTest {
     @Test
     void 금액이_음수이면_예외가_발생한다() {
         // when & then
-        assertThatThrownBy(() -> PurchaseAmountValidator.validate("-1000"))
+        assertThatThrownBy(() -> PurchaseAmountParser.parse("-1000"))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage(ErrorMessage.INVALID_NUMBER.getMessage());
     }
@@ -61,7 +62,7 @@ class PurchaseAmountValidatorTest {
         long unit = lotto.domain.Lotto.PRICE;
 
         //when & then
-        assertThatThrownBy(() -> PurchaseAmountValidator.validate(input))
+        assertThatThrownBy(() -> PurchaseAmountParser.parse(input))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage(ErrorMessage.INVALID_UNIT.getMessage(unit));
     }
@@ -69,7 +70,7 @@ class PurchaseAmountValidatorTest {
     @Test
     @DisplayName("유효한 금액이면 그대로 반환한다")
     void 유효한_금액이면_그대로_반환한다() {
-        long result = PurchaseAmountValidator.validate("8000");
+        long result = PurchaseAmountParser.parse("8000");
         assertThat(result).isEqualTo(8000L);
     }
 }
